+++ slideshowfx-code-extension/src/main/java/com/twasyl/slideshowfx/content/extension/code/CodeContentExtension.java
@@ -25,7 +25,7 @@ import static java.util.regex.Pattern.MULTILINE;
  * This extension supports HTML and Textile markup languages.
  *
  * @author Thierry Wasylczenko
- * @version 1.2
+ * @version 1.3
  * @since SlideshowFX 1.0
  */
 public class CodeContentExtension extends AbstractContentExtension {
@@ -41,7 +41,7 @@ public class CodeContentExtension extends AbstractContentExtension {
                 "Insert code",
                 "Insert code");
 
-        final String baseURL = "prism/1.11.0/";
+        final String baseURL = "prism/1.15.0/";
 
         // Add URL
         this.putResource(ResourceType.CSS_FILE, baseURL.concat("prism.css"));
@@ -68,12 +68,8 @@ public class CodeContentExtension extends AbstractContentExtension {
     public String buildContentString(IMarkup markup) {
         final StringBuilder builder = new StringBuilder();
 
-        if (markup == null || "HTML".equals(markup.getCode())) {
-            builder.append(this.buildDefaultContentString());
-        } else if ("TEXTILE".equals(markup.getCode())) {
+        if ("TEXTILE".equals(markup.getCode()) && !this.controller.shouldHighlightLines()) {
             builder.append(this.buildTextileContentString());
-        } else if ("MARKDOWN".equals(markup.getCode()) && !this.controller.isShowingLineNumbers()) {
-            builder.append(this.buildMarkdownContentString());
         } else {
             builder.append(this.buildDefaultContentString());
         }
@@ -84,7 +80,13 @@ public class CodeContentExtension extends AbstractContentExtension {
     @Override
     public String buildDefaultContentString() {
         final StringBuilder builder = new StringBuilder();
-        builder.append("<pre").append(this.buildDefaultCssClass()).append("><code").append(this.buildDefaultCssClass())
+        builder.append("<pre").append(this.buildDefaultCssClass());
+
+        if (this.controller.shouldHighlightLines()) {
+            builder.append(" data-line=\"").append(this.controller.getHightlightedLines()).append("\"");
+        }
+
+        builder.append("><code").append(this.buildDefaultCssClass())
                 .append(">")
                 .append(this.controller.getCode())
                 .append("</code></pre>");
@@ -121,18 +123,6 @@ public class CodeContentExtension extends AbstractContentExtension {
         if (this.controller.isShowingLineNumbers()) cssClass.add(LINE_NUMBERS_CSS_CLASS);
 
         return cssClass.length() == 2 ? "" : cssClass.toString();
-    }
-
-    protected String buildMarkdownContentString() {
-        final StringBuilder builder = new StringBuilder("```");
-
-        if (this.controller.getLanguage() != null) {
-            builder.append(this.controller.getLanguage().getCssClass());
-        }
-
-        builder.append("\n").append(this.controller.getCode()).append("\n```");
-
-        return builder.toString();
     }
 
     protected boolean codeContainsBlankLines(final String code) {